Add SIMD for loongarch
1. Add VAACalcSadBgd_lasx.
2. Refine WelsSampleSad8x8x2_lasx & WelsIDctT4Rec_lasx & WelsIDctFourT4Rec_lasx.

+++ test/encoder/EncUT_DecodeMbAux.cpp
@@ -253,7 +253,7 @@ TEST (DecodeMbAuxTest, WelsIDctT4Rec_mmi) {
 #endif
 #if defined(HAVE_LASX)
 TEST (DecodeMbAuxTest, WelsIDctT4Rec_lasx) {
-  TestIDctT4Rec<int32_t> (WelsIDctT4Rec_lasx);
+  TestIDctT4Rec<int16_t> (WelsIDctT4Rec_lasx);
 }
 #endif
 template<typename clip_t>
@@ -428,6 +428,6 @@ TEST (DecodeMbAuxTest, WelsIDctRecI16x16Dc_mmi) {
 #endif
 #if defined(HAVE_LASX)
 TEST (DecodeMbAuxTest, WelsIDctFourT4Rec_lasx) {
-  TestIDctFourT4Rec<int32_t> (WelsIDctFourT4Rec_lasx);
+  TestIDctFourT4Rec<int16_t> (WelsIDctFourT4Rec_lasx);
 }
 #endif
